+++ src/test/java/com/hotels/hcommon/hive/metastore/iterator/PartitionIteratorTest.java
@@ -33,6 +33,8 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
+import com.hotels.hcommon.hive.metastore.iterator.PartitionIterator.Ordering;
+
 @RunWith(MockitoJUnitRunner.class)
 public class PartitionIteratorTest {
 
@@ -43,16 +45,16 @@ public class PartitionIteratorTest {
   @Mock
   private Table table;
 
-  private final Partition p20120101 = new Partition();
-  private final Partition p20120102 = new Partition();
-  private final Partition p20120103 = new Partition();
-  private final Partition p20120104 = new Partition();
-  private final Partition p20131009 = new Partition();
-  private final Partition p20131011 = new Partition();
-  private final Partition p20140305 = new Partition();
-  private final Partition p20140324 = new Partition();
-  private final Partition p20140513 = new Partition();
-  private final Partition p20160804 = new Partition();
+  private final Partition p20120101 = newPartition("2012", "01", "01");
+  private final Partition p20120102 = newPartition("2012", "01", "02");
+  private final Partition p20120103 = newPartition("2012", "01", "03");
+  private final Partition p20120104 = newPartition("2012", "01", "04");
+  private final Partition p20131009 = newPartition("2013", "10", "09");
+  private final Partition p20131011 = newPartition("2013", "10", "11");
+  private final Partition p20140305 = newPartition("2014", "03", "05");
+  private final Partition p20140324 = newPartition("2014", "03", "24");
+  private final Partition p20140513 = newPartition("2014", "05", "13");
+  private final Partition p20160804 = newPartition("2016", "08", "04");
 
   @Before
   public void initMocks() throws MetaException, TException {
@@ -60,23 +62,27 @@ public class PartitionIteratorTest {
     when(table.getTableName()).thenReturn(TABLE_NAME);
 
     when(metastore.listPartitionNames(DATABASE_NAME, TABLE_NAME, (short) -1))
-        .thenReturn(Arrays.asList("year=2012/month=01/day=01", "year=2012/month=01/day=02", "year=2012/month=01/day=03",
-            "year=2012/month=01/day=04", "year=2013/month=10/day=09", "year=2013/month=10/day=11",
-            "year=2014/month=03/day=05", "year=2014/month=03/day=24", "year=2014/month=05/day=13",
-            "year=2016/month=08/day=04"));
+        .thenReturn(Arrays
+            .asList("year=2012/month=01/day=01", "year=2012/month=01/day=02", "year=2012/month=01/day=03",
+                "year=2012/month=01/day=04", "year=2013/month=10/day=09", "year=2013/month=10/day=11",
+                "year=2014/month=03/day=05", "year=2014/month=03/day=24", "year=2014/month=05/day=13",
+                "year=2016/month=08/day=04"));
   }
 
   @Test
   public void batching() throws MetaException, TException {
-    when(metastore.getPartitionsByNames(DATABASE_NAME, TABLE_NAME,
-        Arrays.asList("year=2012/month=01/day=01", "year=2012/month=01/day=02", "year=2012/month=01/day=03")))
-            .thenReturn(Arrays.asList(p20120101, p20120102, p20120103));
-    when(metastore.getPartitionsByNames(DATABASE_NAME, TABLE_NAME,
-        Arrays.asList("year=2012/month=01/day=04", "year=2013/month=10/day=09", "year=2013/month=10/day=11")))
-            .thenReturn(Arrays.asList(p20120104, p20131009, p20131011));
-    when(metastore.getPartitionsByNames(DATABASE_NAME, TABLE_NAME,
-        Arrays.asList("year=2014/month=03/day=05", "year=2014/month=03/day=24", "year=2014/month=05/day=13")))
-            .thenReturn(Arrays.asList(p20140305, p20140324, p20140513));
+    when(metastore
+        .getPartitionsByNames(DATABASE_NAME, TABLE_NAME,
+            Arrays.asList("year=2012/month=01/day=01", "year=2012/month=01/day=02", "year=2012/month=01/day=03")))
+                .thenReturn(Arrays.asList(p20120101, p20120102, p20120103));
+    when(metastore
+        .getPartitionsByNames(DATABASE_NAME, TABLE_NAME,
+            Arrays.asList("year=2012/month=01/day=04", "year=2013/month=10/day=09", "year=2013/month=10/day=11")))
+                .thenReturn(Arrays.asList(p20120104, p20131009, p20131011));
+    when(metastore
+        .getPartitionsByNames(DATABASE_NAME, TABLE_NAME,
+            Arrays.asList("year=2014/month=03/day=05", "year=2014/month=03/day=24", "year=2014/month=05/day=13")))
+                .thenReturn(Arrays.asList(p20140305, p20140324, p20140513));
     when(metastore.getPartitionsByNames(DATABASE_NAME, TABLE_NAME, Arrays.asList("year=2016/month=08/day=04")))
         .thenReturn(Arrays.asList(p20160804));
 
@@ -106,13 +112,16 @@ public class PartitionIteratorTest {
 
   @Test
   public void noBatchSmallerThanBatchSize() throws MetaException, TException {
-    when(metastore.getPartitionsByNames(DATABASE_NAME, TABLE_NAME,
-        Arrays.asList("year=2012/month=01/day=01", "year=2012/month=01/day=02", "year=2012/month=01/day=03",
-            "year=2012/month=01/day=04", "year=2013/month=10/day=09", "year=2013/month=10/day=11",
-            "year=2014/month=03/day=05", "year=2014/month=03/day=24", "year=2014/month=05/day=13",
-            "year=2016/month=08/day=04")))
-                .thenReturn(Arrays.asList(p20120101, p20120102, p20120103, p20120104, p20131009, p20131011, p20140305,
-                    p20140324, p20140513, p20160804));
+    when(metastore
+        .getPartitionsByNames(DATABASE_NAME, TABLE_NAME,
+            Arrays
+                .asList("year=2012/month=01/day=01", "year=2012/month=01/day=02", "year=2012/month=01/day=03",
+                    "year=2012/month=01/day=04", "year=2013/month=10/day=09", "year=2013/month=10/day=11",
+                    "year=2014/month=03/day=05", "year=2014/month=03/day=24", "year=2014/month=05/day=13",
+                    "year=2016/month=08/day=04")))
+                        .thenReturn(Arrays
+                            .asList(p20120101, p20120102, p20120103, p20120104, p20131009, p20131011, p20140305,
+                                p20140324, p20140513, p20160804));
 
     PartitionIterator iterator = new PartitionIterator(metastore, table, (short) 11);
     assertThat(iterator.hasNext(), is(true));
@@ -147,4 +156,51 @@ public class PartitionIteratorTest {
     assertThat(iterator.hasNext(), is(false));
   }
 
+  @Test
+  public void reversOrderWithbatching() throws MetaException, TException {
+    when(metastore
+        .getPartitionsByNames(DATABASE_NAME, TABLE_NAME,
+            Arrays.asList("year=2016/month=08/day=04", "year=2014/month=05/day=13", "year=2014/month=03/day=24")))
+                .thenReturn(Arrays.asList(p20140324, p20140513, p20160804));
+    when(metastore
+        .getPartitionsByNames(DATABASE_NAME, TABLE_NAME,
+            Arrays.asList("year=2014/month=03/day=05", "year=2013/month=10/day=11", "year=2013/month=10/day=09")))
+                .thenReturn(Arrays.asList(p20131009, p20131011, p20140305));
+    when(metastore
+        .getPartitionsByNames(DATABASE_NAME, TABLE_NAME,
+            Arrays.asList("year=2012/month=01/day=04", "year=2012/month=01/day=03", "year=2012/month=01/day=02")))
+                .thenReturn(Arrays.asList(p20120102, p20120103, p20120104));
+
+    when(metastore.getPartitionsByNames(DATABASE_NAME, TABLE_NAME, Arrays.asList("year=2012/month=01/day=01")))
+        .thenReturn(Arrays.asList(p20120101));
+
+    PartitionIterator iterator = new PartitionIterator(metastore, table, (short) 3, Ordering.REVERSE);
+    assertThat(iterator.hasNext(), is(true));
+    assertThat(iterator.next(), is(p20160804));
+    assertThat(iterator.hasNext(), is(true));
+    assertThat(iterator.next(), is(p20140513));
+    assertThat(iterator.hasNext(), is(true));
+    assertThat(iterator.next(), is(p20140324));
+    assertThat(iterator.hasNext(), is(true));
+    assertThat(iterator.next(), is(p20140305));
+    assertThat(iterator.hasNext(), is(true));
+    assertThat(iterator.next(), is(p20131011));
+    assertThat(iterator.hasNext(), is(true));
+    assertThat(iterator.next(), is(p20131009));
+    assertThat(iterator.hasNext(), is(true));
+    assertThat(iterator.next(), is(p20120104));
+    assertThat(iterator.hasNext(), is(true));
+    assertThat(iterator.next(), is(p20120103));
+    assertThat(iterator.hasNext(), is(true));
+    assertThat(iterator.next(), is(p20120102));
+    assertThat(iterator.hasNext(), is(true));
+    assertThat(iterator.next(), is(p20120101));
+    assertThat(iterator.hasNext(), is(false));
+  }
+
+  private Partition newPartition(String year, String month, String day) {
+    Partition result = new Partition();
+    result.setValues(Arrays.asList(year, month, day));
+    return result;
+  }
 }