+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # [TBD] - TBD
 ### Changed
 * Upgraded hotels-oss-parent to 2.3.3 (was 2.1.0).
+### Added
+* Made PartitionIterator optionally run in reverse order.
 
 # [1.2.0] - 2018-09-21
 ### Added